requirements.txt
tortoise-orm
fastapi
pydantic
uvicorn